Split results into separate files.
